requirements.txt
biopython
tabulate
